Move generator commands
